microservicios/requirements.txt
django == 4.1.7
django-rest-framework